chore: trade input tab managed by the header instead of prop drilling
