+++ src/components/MultiHopTrade/components/SharedTradeInput/SharedTradeInputHeader.tsx
@@ -1,35 +1,51 @@
 import { CardHeader, Flex, Heading } from '@chakra-ui/react'
+import { assertUnreachable } from '@shapeshiftoss/utils'
 import { useCallback, useState } from 'react'
 import { useTranslate } from 'react-polyglot'
+import { useHistory } from 'react-router'
 import { useFeatureFlag } from 'hooks/useFeatureFlag/useFeatureFlag'
 
-import { TradeInputTab } from '../../types'
+import { TradeInputTab, TradeRoutePaths } from '../../types'
 
 type SharedTradeInputHeaderProps = {
   initialTab: TradeInputTab
   rightContent?: JSX.Element
-  onChangeTab: (newTab: TradeInputTab) => void
 }
 
 export const SharedTradeInputHeader = ({
   initialTab,
   rightContent,
-  onChangeTab,
 }: SharedTradeInputHeaderProps) => {
+  const history = useHistory()
   const translate = useTranslate()
   const [selectedTab, setSelectedTab] = useState<TradeInputTab>(initialTab)
 
   const enableBridgeClaims = useFeatureFlag('ArbitrumBridgeClaims')
 
+  const handleChangeTab = useCallback(
+    (newTab: TradeInputTab) => {
+      setSelectedTab(newTab)
+      switch (newTab) {
+        case TradeInputTab.Trade:
+          history.push(TradeRoutePaths.Input)
+          break
+        case TradeInputTab.Claim:
+          history.push(TradeRoutePaths.Claim)
+          break
+        default:
+          assertUnreachable(newTab)
+      }
+    },
+    [history],
+  )
+
   const handleClickTrade = useCallback(() => {
-    setSelectedTab(TradeInputTab.Trade)
-    onChangeTab(TradeInputTab.Trade)
-  }, [onChangeTab])
+    handleChangeTab(TradeInputTab.Trade)
+  }, [handleChangeTab])
 
   const handleClickClaim = useCallback(() => {
-    setSelectedTab(TradeInputTab.Claim)
-    onChangeTab(TradeInputTab.Claim)
-  }, [onChangeTab])
+    handleChangeTab(TradeInputTab.Claim)
+  }, [handleChangeTab])
 
   return (
     <CardHeader px={6}>
